Сделал связь между профессором и ассистентом, Добавил file storage

+++ diploma.client/src/app/pages/my-projects/my-projects.component.ts
@@ -5,7 +5,7 @@ import {MyProjectCard} from "../../../model/MyProjectCard";
 import {MyProjectsService} from "./my-projects.service";
 import {MatDialog} from "@angular/material";
 import {MyProjectEditDialogComponent} from "./components/my-project-edit-dialog/my-project-edit-dialog.component";
-import {MyProjectDetail} from "../../../model/MyProjectDetail";
+import {PopupAskComponent} from "../../common/popup-ask/popup-ask.component";
 
 @Component({
   selector: 'app-my-projects',
@@ -26,7 +26,10 @@ export class MyProjectsComponent implements OnInit {
 
   ngOnInit() {
     this.menuListService.menuList.then(value => this.menuItems = value);
-    this.myProjectsService.myProjects.then(value => this.myProjects = value);
+    this.myProjectsService.myProjects.then(value => {
+      console.log(value);
+      this.myProjects = value
+    });
   }
 
   editMyProject(id: string) {
@@ -49,10 +52,15 @@ export class MyProjectsComponent implements OnInit {
     let myProject = this.myProjects.find(value => value.id === id);
     myProject.isButtonsDisabled = true;
 
-    this.myProjectsService.deleteMyProject(myProject.id)
-      .then(() => {
-        this.myProjects = this.myProjects.filter(value => value.id !== id);
-      });
+    const matDialogRef = this.matDialog.open(PopupAskComponent, {data: 'Are you sure?'});
+    matDialogRef.afterClosed().subscribe(value => {
+      myProject.isButtonsDisabled = false;
+      if (!value) return;
+      this.myProjectsService.deleteMyProject(myProject.id)
+        .then(() => {
+          this.myProjects = this.myProjects.filter(value => value.id !== id);
+        });
+    });
 
   }
 }
